2016/04 Security Through Obscurity/aoc_201604_test.py
import pytest
import aoc_201604 as aoc


@pytest.mark.parametrize(
    "puzzle_input,expected,extra",
    [
        (
            "aaaaa-bbb-z-y-x-123[abxyz]\na-b-c-d-e-f-g-h-987[abcde]\nnot-a-real-room-404[oarel]\ntotally-real-room-200[decoy]",
            "1514",
            None,
        )
    ],
)
def test_part_a(puzzle_input, expected, extra):
    assert aoc.solve_a(puzzle_input) == expected
    # assert aoc.solve(puzzle_input, int(extra.split("=")[1])) == expected


@pytest.mark.parametrize(
    "puzzle_input,expected,extra",
    [("qzmt-zixmtkozy-ivhz-343[zimth]", "very encrypted name", "")],
)
def test_part_b(puzzle_input, expected, extra):
    assert aoc.solve_b(puzzle_input) == expected
    # assert aoc.solve(puzzle_input, int(extra.split("=")[1]), is_b=True) == expected
